move PortalTarget._dependOnScope to PortalLinkScope.of
