+++ lib/src/portal_link.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import '../flutter_portal.dart';
@@ -61,4 +62,56 @@ class PortalLinkScope extends InheritedWidget {
   }
 
   bool linkEquals(PortalLinkScope other) => portalLink == other.portalLink;
+
+  static PortalLinkScope? of(
+    BuildContext context,
+    List<PortalLabel<dynamic>> portalCandidateLabels,
+  ) {
+    for (final portalLabel in portalCandidateLabels) {
+      final scope =
+          context.dependOnSpecificInheritedWidgetOfExactType<PortalLinkScope>(
+              (scope) => scope.portalLabels.contains(portalLabel));
+      if (scope != null) {
+        return scope;
+      }
+    }
+    return null;
+  }
+}
+
+extension BuildContextPortalLinkScopeExt on BuildContext {
+  /// https://stackoverflow.com/questions/71200969
+  Iterable<InheritedElement> getElementsForInheritedWidgetsOfExactType<
+      T extends InheritedWidget>() sync* {
+    final element = getElementForInheritedWidgetOfExactType<T>();
+    if (element != null) {
+      yield element;
+
+      Element? parent;
+      element.visitAncestorElements((element) {
+        parent = element;
+        return false;
+      });
+
+      if (parent != null) {
+        yield* parent!.getElementsForInheritedWidgetsOfExactType<T>();
+      }
+    }
+  }
+
+  InheritedElement? getSpecificElementForInheritedWidgetsOfExactType<
+          T extends InheritedWidget>(bool Function(T) test) =>
+      getElementsForInheritedWidgetsOfExactType<T>()
+          .where((element) => test(element.widget as T))
+          .firstOrNull;
+
+  /// https://stackoverflow.com/questions/71200969
+  T? dependOnSpecificInheritedWidgetOfExactType<T extends InheritedWidget>(
+      bool Function(T) test) {
+    final element = getSpecificElementForInheritedWidgetsOfExactType<T>(test);
+    if (element == null) {
+      return null;
+    }
+    return dependOnInheritedElement(element) as T;
+  }
 }
